moved file as requested and corrected json translate file

Related Issue: SPBK-4003
