+++ ui/src/locales/en/translation.json
@@ -319,7 +319,7 @@
   "modal.rule.title": "Create a New Rule",
   "no.data": "No Data",
   "none": "None",
-  "notebook.opensPanel": "Opens notebook panel",
+  "notebook.opens.panel": "Opens notebook panel",
   "notebook.title": "Notebook",
   "on": "on",
   "open": "Open",
@@ -426,7 +426,7 @@
   "personalization.reset_text": "Reset to site defaults",
   "personalization.showbreadcrumbs": "Show Breadcrumbs",
   "personalization.sticky": "Sticky Topbar",
-  "pivot.noDescription": "No description",
+  "pivot.description.none": "No description",
   "pivot.openDossier": "Open Dossier",
   "pivot.url": "URL:",
   "query": "Query",
