+++ ui/src/locales/fr/translation.json
@@ -319,7 +319,7 @@
   "modal.rule.title": "Créer une nouvelle règle",
   "no.data": "Aucune donnée",
   "none": "Rien",
-  "notebook.opensPanel": "Ouvre le panneau des notes",
+  "notebook.opens.panel": "Ouvre le panneau des notes",
   "notebook.title": "Carnet",
   "on": "sur",
   "open": "Ouvert",
@@ -426,7 +426,7 @@
   "personalization.reset_text": "Réinitialiser les paramètres",
   "personalization.showbreadcrumbs": "Afficher fils d'ariane",
   "personalization.sticky": "Barre supérieure collante",
-  "pivot.noDescription": "Aucune description",
+  "pivot.description.none": "Aucune description",
   "pivot.openDossier": "Ouvrir le dossier",
   "pivot.url": "URL :",
   "query": "Requête",
